started thinking about why I do not include f after constant floats in blendelf, and came to the conclusion that it might not be a that good idea after all... changed all consts to include the f

+++ elf/light.h
@@ -10,30 +10,30 @@ ELF_API elfLight* ELF_APIENTRY elfCreateLight(const char* name)
 
 	elfInitActor((elfActor*)light, ELF_FALSE);
 
-	light->distance = 30.0;
-	light->fadeSpeed = 1.0/30.0;
-	light->innerCone = 45.0;
-	light->outerCone = 0.0;
+	light->distance = 30.0f;
+	light->fadeSpeed = 1.0f/30.0f;
+	light->innerCone = 45.0f;
+	light->outerCone = 0.0f;
 	light->lightType = ELF_POINT_LIGHT;
 	light->visible = ELF_TRUE;
 	light->shaft = ELF_FALSE;
-	light->shaftSize = 1.0;
-	light->shaftIntensity = 1.0;
-	light->shaftFadeOff = 0.0;
+	light->shaftSize = 1.0f;
+	light->shaftIntensity = 1.0f;
+	light->shaftFadeOff = 0.0f;
 
-	elfSetLightColor(light, 1.0, 1.0, 1.0, 1.0);
+	elfSetLightColor(light, 1.0f, 1.0f, 1.0f, 1.0f);
 
 	light->shadowCamera = elfCreateCamera("elfShadowCamera");
 	elfSetCameraViewport(light->shadowCamera, 0, 0, 512, 512);
-	elfSetCameraPerspective(light->shadowCamera, (light->innerCone+light->outerCone)*2, 1.0, 1.0, light->distance+(1.0/light->fadeSpeed));
+	elfSetCameraPerspective(light->shadowCamera, (light->innerCone+light->outerCone)*2, 1.0f, 1.0f, light->distance+(1.0f/light->fadeSpeed));
 
 	gfxMatrix4SetIdentity(light->projectionMatrix);
 
-	light->dobject = elfCreatePhysicsObjectSphere(0.5, 0.0, 0.0, 0.0, 0.0);
+	light->dobject = elfCreatePhysicsObjectSphere(0.5f, 0.0f, 0.0f, 0.0f, 0.0f);
 	elfSetPhysicsObjectActor(light->dobject, (elfActor*)light);
 	elfIncRef((elfObject*)light->dobject);
 
-	light->pbbLengths.x = light->pbbLengths.y = light->pbbLengths.z = 1.0;
+	light->pbbLengths.x = light->pbbLengths.y = light->pbbLengths.z = 1.0f;
 
 	if(name) light->name = elfCreateString(name);
 
@@ -91,16 +91,16 @@ ELF_API void ELF_APIENTRY elfSetLightColor(elfLight* light, float r, float g, fl
 ELF_API void ELF_APIENTRY elfSetLightDistance(elfLight* light, float distance)
 {
 	light->distance = distance;
-	if(light->distance < 0.0) light->distance = 0.0;
-	elfSetCameraPerspective(light->shadowCamera, (light->innerCone+light->outerCone)*2, 1.0, 1.0, light->distance+(1.0/(light->fadeSpeed)));
+	if(light->distance < 0.0f) light->distance = 0.0f;
+	elfSetCameraPerspective(light->shadowCamera, (light->innerCone+light->outerCone)*2, 1.0f, 1.0f, light->distance+(1.0f/(light->fadeSpeed)));
 }
 
 ELF_API void ELF_APIENTRY elfSetLightFadeSpeed(elfLight* light, float fadeSpeed)
 {
 	light->fadeSpeed = fadeSpeed;
-	if(light->fadeSpeed < 0.0001) light->fadeSpeed = 0.0001;
-	if(light->fadeSpeed > 1.0) light->fadeSpeed = 1.0;
-	elfSetCameraPerspective(light->shadowCamera, (light->innerCone+light->outerCone)*2, 1.0, 1.0, light->distance+(1.0/(light->fadeSpeed)));
+	if(light->fadeSpeed < 0.0001f) light->fadeSpeed = 0.0001f;
+	if(light->fadeSpeed > 1.0f) light->fadeSpeed = 1.0f;
+	elfSetCameraPerspective(light->shadowCamera, (light->innerCone+light->outerCone)*2, 1.0f, 1.0f, light->distance+(1.0f/(light->fadeSpeed)));
 }
 
 ELF_API void ELF_APIENTRY elfSetLightShadows(elfLight* light, unsigned char shadows)
@@ -181,9 +181,9 @@ ELF_API void ELF_APIENTRY elfSetLightCone(elfLight* light, float innerCone, floa
 {
 	light->innerCone = innerCone;
 	light->outerCone = outerCone;
-	if(light->innerCone < 0.0) light->innerCone = 0.0;
-	if(light->outerCone < 0.0) light->outerCone = 0.0;
-	elfSetCameraPerspective(light->shadowCamera, (light->innerCone+light->outerCone)*2, 1.0, 1.0, light->distance+(1.0/(light->fadeSpeed)));
+	if(light->innerCone < 0.0f) light->innerCone = 0.0f;
+	if(light->outerCone < 0.0f) light->outerCone = 0.0f;
+	elfSetCameraPerspective(light->shadowCamera, (light->innerCone+light->outerCone)*2, 1.0f, 1.0f, light->distance+(1.0f/(light->fadeSpeed)));
 }
 
 ELF_API void ELF_APIENTRY elfSetLightShaft(elfLight* light, float size, float intensity, float fadeOff)
@@ -191,10 +191,10 @@ ELF_API void ELF_APIENTRY elfSetLightShaft(elfLight* light, float size, float in
 	light->shaft = ELF_TRUE;
 	light->shaftSize = size;
 	light->shaftIntensity = intensity;
-	if(light->shaftIntensity < 0.0) light->shaftIntensity = 0.0;
+	if(light->shaftIntensity < 0.0f) light->shaftIntensity = 0.0f;
 	light->shaftFadeOff = fadeOff;
-	if(light->shaftFadeOff < 0.0) light->shaftFadeOff = 0.0;
-	if(light->shaftFadeOff > 1.0) light->shaftFadeOff = 1.0;
+	if(light->shaftFadeOff < 0.0f) light->shaftFadeOff = 0.0f;
+	if(light->shaftFadeOff > 1.0f) light->shaftFadeOff = 1.0f;
 }
 
 ELF_API void ELF_APIENTRY elfDisableLightShaft(elfLight* light)
@@ -205,7 +205,7 @@ ELF_API void ELF_APIENTRY elfDisableLightShaft(elfLight* light)
 void elfSetLight(elfLight* light, elfCamera* camera, gfxShaderParams* shaderParams)
 {
 	float axis[3];
-	float finalAxis[3] = {0.0, 0.0, -1.0};
+	float finalAxis[3] = {0.0f, 0.0f, -1.0f};
 	float matrix[16];
 	float matrix2[16];
 	float orient[4];
@@ -262,98 +262,98 @@ void elfDrawLightDebug(elfLight* light, gfxShaderParams* shaderParams)
 
 	vertexBuffer = (float*)gfxGetVertexDataBuffer(eng->lines);
 
-	if(!light->selected) gfxSetColor(&shaderParams->materialParams.diffuseColor, 0.5, 0.5, 0.2, 1.0);
-	else gfxSetColor(&shaderParams->materialParams.diffuseColor, 1.0, 0.0, 0.0, 1.0);
+	if(!light->selected) gfxSetColor(&shaderParams->materialParams.diffuseColor, 0.5f, 0.5f, 0.2f, 1.0f);
+	else gfxSetColor(&shaderParams->materialParams.diffuseColor, 1.0f, 0.0f, 0.0f, 1.0f);
 	gfxSetShaderParams(shaderParams);
 
-	step = (360.0/((float)32))*GFX_PI_DIV_180;
+	step = (360.0f/32.0)*GFX_PI_DIV_180;
 
 	for(i = 0; i < 32; i++)
 	{
-		vertexBuffer[i*3] = -((float)sin((float)(step*i)))*0.5;
-		vertexBuffer[i*3+1] = ((float)cos((float)(step*i)))*0.5;
-		vertexBuffer[i*3+2] = 0.0;
+		vertexBuffer[i*3] = -((float)sin((float)(step*i)))*0.5f;
+		vertexBuffer[i*3+1] = ((float)cos((float)(step*i)))*0.5f;
+		vertexBuffer[i*3+2] = 0.0f;
 	}
 
 	gfxDrawLineLoop(32, eng->lines);
 
 	for(i = 0; i < 32; i++)
 	{
-		vertexBuffer[i*3] = 0.0;
-		vertexBuffer[i*3+1] = -((float)sin((float)(step*i)))*0.5;
-		vertexBuffer[i*3+2] = ((float)cos((float)(step*i)))*0.5;
+		vertexBuffer[i*3] = 0.0f;
+		vertexBuffer[i*3+1] = -((float)sin(step*i))*0.5f;
+		vertexBuffer[i*3+2] = ((float)cos(step*i))*0.5f;
 	}
 
 	gfxDrawLineLoop(32, eng->lines);
 
 	for(i = 0; i < 32; i++)
 	{
-		vertexBuffer[i*3] = -((float)sin((float)(step*i)))*0.5;
-		vertexBuffer[i*3+1] = 0.0;
-		vertexBuffer[i*3+2] = ((float)cos((float)(step*i)))*0.5;
+		vertexBuffer[i*3] = -((float)sin(step*i))*0.5f;
+		vertexBuffer[i*3+1] = 0.0f;
+		vertexBuffer[i*3+2] = ((float)cos(step*i))*0.5f;
 	}
 
 	gfxDrawLineLoop(32, eng->lines);
 
-	if(!light->selected) gfxSetColor(&shaderParams->materialParams.diffuseColor, 0.3, 0.3, 0.1, 1.0);
-	else gfxSetColor(&shaderParams->materialParams.diffuseColor, 1.0, 0.0, 0.0, 1.0);
+	if(!light->selected) gfxSetColor(&shaderParams->materialParams.diffuseColor, 0.3f, 0.3f, 0.1f, 1.0f);
+	else gfxSetColor(&shaderParams->materialParams.diffuseColor, 1.0f, 0.0f, 0.0f, 1.0f);
 	gfxSetShaderParams(shaderParams);
 
 	if(light->lightType == ELF_POINT_LIGHT)
 	{
-		step = (360.0/((float)128))*GFX_PI_DIV_180;
+		step = (360.0f/128.0f)*GFX_PI_DIV_180;
 
 		for(i = 0; i < 128; i++)
 		{
-			vertexBuffer[i*3] = -((float)sin((float)(step*i)))*light->distance;
-			vertexBuffer[i*3+1] = ((float)cos((float)(step*i)))*light->distance;
-			vertexBuffer[i*3+2] = 0.0;
+			vertexBuffer[i*3] = -((float)sin(step*i))*light->distance;
+			vertexBuffer[i*3+1] = ((float)cos(step*i))*light->distance;
+			vertexBuffer[i*3+2] = 0.0f;
 		}
 
 		gfxDrawLines(128, eng->lines);
 
 		for(i = 0; i < 128; i++)
 		{
-			vertexBuffer[i*3] = 0.0;
-			vertexBuffer[i*3+1] = -((float)sin((float)(step*i)))*light->distance;
-			vertexBuffer[i*3+2] = ((float)cos((float)(step*i)))*light->distance;
+			vertexBuffer[i*3] = 0.0f;
+			vertexBuffer[i*3+1] = -((float)sin(step*i))*light->distance;
+			vertexBuffer[i*3+2] = ((float)cos(step*i))*light->distance;
 		}
 
 		gfxDrawLines(128, eng->lines);
 
 		for(i = 0; i < 128; i++)
 		{
-			vertexBuffer[i*3] = -((float)sin((float)(step*i)))*light->distance;
-			vertexBuffer[i*3+1] = 0.0;
-			vertexBuffer[i*3+2] = ((float)cos((float)(step*i)))*light->distance;
+			vertexBuffer[i*3] = -((float)sin(step*i))*light->distance;
+			vertexBuffer[i*3+1] = 0.0f;
+			vertexBuffer[i*3+2] = ((float)cos(step*i))*light->distance;
 		}
 
 		gfxDrawLines(128, eng->lines);
 
-		vertexBuffer[0] = 0.0;
-		vertexBuffer[1] = 0.0;
-		vertexBuffer[2] = 0.5;
-		vertexBuffer[3] = 0.0;
-		vertexBuffer[4] = 0.0;
-		vertexBuffer[5] = -0.5;
-		vertexBuffer[6] = -0.5;
-		vertexBuffer[7] = 0.0;
-		vertexBuffer[8] = 0.0;
-		vertexBuffer[9] = 0.5;
-		vertexBuffer[10] = 0.0;
-		vertexBuffer[11] = 0.0;
-		vertexBuffer[12] = 0.0;
-		vertexBuffer[13] = -0.5;
-		vertexBuffer[14] = 0.0;
-		vertexBuffer[15] = 0.0;
-		vertexBuffer[16] = 0.5;
-		vertexBuffer[17] = 0.0;
+		vertexBuffer[0] = 0.0f;
+		vertexBuffer[1] = 0.0f;
+		vertexBuffer[2] = 0.5f;
+		vertexBuffer[3] = 0.0f;
+		vertexBuffer[4] = 0.0f;
+		vertexBuffer[5] = -0.5f;
+		vertexBuffer[6] = -0.5f;
+		vertexBuffer[7] = 0.0f;
+		vertexBuffer[8] = 0.0f;
+		vertexBuffer[9] = 0.5f;
+		vertexBuffer[10] = 0.0f;
+		vertexBuffer[11] = 0.0f;
+		vertexBuffer[12] = 0.0f;
+		vertexBuffer[13] = -0.5f;
+		vertexBuffer[14] = 0.0f;
+		vertexBuffer[15] = 0.0f;
+		vertexBuffer[16] = 0.5f;
+		vertexBuffer[17] = 0.0f;
 
 		gfxDrawLines(6, eng->lines);
 	}
 	else if(light->lightType == ELF_SPOT_LIGHT)
 	{
-		step = (360.0/((float)128))*GFX_PI_DIV_180;
+		step = (360.0f/128.0)*GFX_PI_DIV_180;
 		size = sin((float)GFX_PI_DIV_180*light->innerCone)*light->distance;
 
 		for(i = 0; i < 128; i++)
@@ -376,59 +376,59 @@ void elfDrawLightDebug(elfLight* light, gfxShaderParams* shaderParams)
 
 		gfxDrawLines(128, eng->lines);
 
-		vertexBuffer[0] = 0.0;
-		vertexBuffer[1] = 0.0;
-		vertexBuffer[2] = 0.5;
-		vertexBuffer[3] = 0.0;
-		vertexBuffer[4] = 0.0;
-		vertexBuffer[5] = -0.5;
-		vertexBuffer[6] = -0.5;
-		vertexBuffer[7] = 0.0;
-		vertexBuffer[8] = 0.0;
-		vertexBuffer[9] = 0.5;
-		vertexBuffer[10] = 0.0;
-		vertexBuffer[11] = 0.0;
-		vertexBuffer[12] = 0.0;
-		vertexBuffer[13] = -0.5;
-		vertexBuffer[14] = 0.0;
-		vertexBuffer[15] = 0.0;
-		vertexBuffer[16] = 0.5;
-		vertexBuffer[17] = 0.0;
-		vertexBuffer[18] = 0.0;
-		vertexBuffer[19] = 0.0;
-		vertexBuffer[20] = 0.0;
+		vertexBuffer[0] = 0.0f;
+		vertexBuffer[1] = 0.0f;
+		vertexBuffer[2] = 0.5f;
+		vertexBuffer[3] = 0.0f;
+		vertexBuffer[4] = 0.0f;
+		vertexBuffer[5] = -0.5f;
+		vertexBuffer[6] = -0.5f;
+		vertexBuffer[7] = 0.0f;
+		vertexBuffer[8] = 0.0f;
+		vertexBuffer[9] = 0.5f;
+		vertexBuffer[10] = 0.0f;
+		vertexBuffer[11] = 0.0f;
+		vertexBuffer[12] = 0.0f;
+		vertexBuffer[13] = -0.5f;
+		vertexBuffer[14] = 0.0f;
+		vertexBuffer[15] = 0.0f;
+		vertexBuffer[16] = 0.5f;
+		vertexBuffer[17] = 0.0f;
+		vertexBuffer[18] = 0.0f;
+		vertexBuffer[19] = 0.0f;
+		vertexBuffer[20] = 0.0f;
 		vertexBuffer[21] = size;
-		vertexBuffer[22] = 0.0;
+		vertexBuffer[22] = 0.0f;
 		vertexBuffer[23] = -light->distance;
-		vertexBuffer[24] = 0.0;
-		vertexBuffer[25] = 0.0;
-		vertexBuffer[26] = 0.0;
+		vertexBuffer[24] = 0.0f;
+		vertexBuffer[25] = 0.0f;
+		vertexBuffer[26] = 0.0f;
 		vertexBuffer[27] = -size;
-		vertexBuffer[28] = 0.0;
+		vertexBuffer[28] = 0.0f;
 		vertexBuffer[29] = -light->distance;
 
 		gfxDrawLines(13, eng->lines);
 	}
 	else if(light->lightType == ELF_SUN_LIGHT)
 	{
-		vertexBuffer[0] = 0.0;
-		vertexBuffer[1] = 0.0;
-		vertexBuffer[2] = 0.5;
-		vertexBuffer[3] = 0.0;
-		vertexBuffer[4] = 0.0;
+		vertexBuffer[0] = 0.0f;
+		vertexBuffer[1] = 0.0f;
+		vertexBuffer[2] = 0.5f;
+		vertexBuffer[3] = 0.0f;
+		vertexBuffer[4] = 0.0f;
 		vertexBuffer[5] = -light->distance;
-		vertexBuffer[6] = -0.5;
-		vertexBuffer[7] = 0.0;
-		vertexBuffer[8] = 0.0;
-		vertexBuffer[9] = 0.5;
-		vertexBuffer[10] = 0.0;
-		vertexBuffer[11] = 0.0;
-		vertexBuffer[12] = 0.0;
-		vertexBuffer[13] = -0.5;
-		vertexBuffer[14] = 0.0;
-		vertexBuffer[15] = 0.0;
-		vertexBuffer[16] = 0.5;
-		vertexBuffer[17] = 0.0;
+		vertexBuffer[6] = -0.5f;
+		vertexBuffer[7] = 0.0f;
+		vertexBuffer[8] = 0.0f;
+		vertexBuffer[9] = 0.5f;
+		vertexBuffer[10] = 0.0f;
+		vertexBuffer[11] = 0.0f;
+		vertexBuffer[12] = 0.0f;
+		vertexBuffer[13] = -0.5f;
+		vertexBuffer[14] = 0.0f;
+		vertexBuffer[15] = 0.0f;
+		vertexBuffer[16] = 0.5f;
+		vertexBuffer[17] = 0.0f;
 
 		gfxDrawLines(6, eng->lines);
 	}
